got timetraps working and eventSystem transfers
basically eventSystem can now transfer parameters through changeParameter (meaning that you can transfer values and do stuff like addition and exponentiation for variable values). This allows for stuff like a timetrap time controlling the radius of a radiusMult and thus "trapping" radius + more stuff i havent thought of yet! I would include this in the clone as well but you can just change the parameters of the parent beforehand
also timetraps working with a sicc radial gradeint in the middle